+++ src/ssh.rs
@@ -1,7 +1,7 @@
 use crate::cfg::{Auth, Host};
 
 use anyhow::{bail, Context, Result};
-use indicatif::{ProgressBar, ProgressStyle};
+use indicatif::ProgressBar;
 use log::{debug, error, info};
 use rpassword::prompt_password_stderr;
 use ssh2::Session as RawSession;
@@ -215,9 +215,7 @@ impl<'a> SshSession<'a> {
             .with_context(|| format!("Could not create remote file: {}", path_remote.display()))?;
 
         let bar = ProgressBar::new(local_file.metadata()?.len());
-        bar.set_style(ProgressStyle::default_bar()
-        .template("{spinner:.green} [{elapsed_precise}] [{bar:40.cyan/blue}] {bytes} / {total_bytes} @ {bytes_per_sec} ({eta})")
-        .progress_chars("#>-"));
+        bar.set_style(crate::cli::style_progress_bar());
         let mut reader = BufReader::new(local_file);
 
         loop {